Create logic for setting the operator on button click
This commit creates the function for recording the operator the user clicks on. It is defined within the 'arithmeticOperation' object to more easily manipulate the 'operator' property within the same object. 

In addition, the function calls for 'addNumbersToDisplay' and 'addOperatorToDisplay' were removed, and the definitions commented out, so they can be refactored in the future solely using JavaScript data structures.