Create logic for setting the operands on button clicks
This commit creates the function for building the record of the operand when the user clicks on the number buttons on the calculator. It is defined within the 'arithmeticOperation' object to more easily manipulate the 'operand1' and 'operand2' properties.

In addition, the values in the 'operand1' and 'operand2' properties were changed from 'null' to an empty string to allow for easy concatenations when a number is clicked.