Create logic for clearing and adding operation result to display
This commit ensures the display only shows the current operand value written into and clears the display when appropriate. When all 'arithmeticOperation' properties contain necessary values and an operator button is clicked, the operation results are displayed. In the future, this will also clear the properties and assign the result to 'operand1'.