Refactor logic for updating the display
This commit cuts down on the amount of code by reworking how the display
is updated. The values to be displayed are now read directly from the
'arithmeticOperation' object and updated whenever a button is clicked.

The operator's empty value is now an empty string rather than null so
that its properly displayed.