Refactor updateDisplay to use getOperationString method
Previously, the `updateDisplay` function used a template literal to
display the arithmetic operation, which was repetitive. To improve the
codebase, we created a new method called `getOperationString` in the
`arithmeticOperation` object, which returns the same string. We then
refactored the `updateDisplayOnOperatorClick`, `updateDisplayOnNumberClick`,
and `updateDisplayOnUtilityClick` functions to use the `getOperationString`
method instead.