Split addCharToOperand into operand-specific methods
Previously, the addCharToOperand method handled adding characters to
both operands depending on whether the operation had a value. For
simplicity, we decided that the method should not handle deciding what
operand to add to. Instead, an event handler outside the ArithmeticOperation
class should handle this. We felt that consistency is important, and no
other method within the class handles both operands.

In its place, we now have two methods called addCharToOperand1 and
addCharToOperand2, which do as the name suggests.