Refactor elementIds to CalculatorUI class
I have refactored the previous 'elementIds' object into a new class
called CalculatorUI, which has improved the organization of the codebase.
The new name more accurately reflects the purpose of the elements.

The conversion to a class has enabled the use of private properties. As
the data within the class is sensitive and should not be modified, all
properties have been made private. I have created only "get" methods to
retrieve the information.

To make it easier for "get" methods to obtain the properties, I have
'flattened out' the properties.

Furthermore, I have updated the code throughout main.js to use the new
CalculatorUI class wherever applicable.