Disable decimal button if operand has decimal
I implemented a new feature to disable the decimal button in case an
operand already contains a decimal point. This functionality was
necessary because a number cannot have more than one decimal point.

To implement this feature, I removed the decimal handling code from the
event listener in the setOperandOnClick function and created a new
function specifically to handle the decimal button. This was done
because there was additional logic required to determine if an operand
can take a decimal point, which is outside the scope of the
setOperandOnClick function.

In addition, I created two new functions within the ArithmeticOperation
class to support the new functionality.